pyproject.toml
[tool.poetry]
name = "barista-order-service"
version = "0.1.0"
description = ""
authors = ["Jeffrey Wu"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pika = "^1.3.2"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.1.1"

httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
